requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.20
django-cors-headers==4.4.0
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.10
itypes==1.2.0
Jinja2==3.1.5
MarkupSafe==2.1.5
packaging==24.2
PyYAML==6.0.2
requests==2.32.3
setuptools==75.3.0
sqlparse==0.5.3
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.2.3